Fix SubSource auto_range_limit
The `SubSource` `auto_range_limit` attribute was coded much like the `SubSense` `auto_range_limit` such that a lower an upper bound are returned/given when queried/written.  However, the Source SCPI specification only gives/takes a lower limit for this command.  This commit fixes this non-working API command by removing the query/write for the upper limit.  This simplifies the implementation by removing the need for a `@property` and `@auto_range_limit.setter` for a simple `b2900._command` call.